requirements.txt
aiogram==3.26.0
aiosqlite==0.21.0
python-dotenv==1.0.1
tzdata